Chapter 6.1 parts 1 through 3
These are to be run alongside networkx and matplotlib. Yes, in part 3, I did do an import for json because the program can write adjacency matrix data to a json file, but as far as the actual problems go, I chose to not go the route of chegg and answer all the homework problems. I followed the chapter topics outlined in the requirements doc.